Moved RequestHandlers and FiniteStateMachine to new folders
