+++ src/main/java/org/icatproject/ids/IdsService.java
@@ -38,11 +38,11 @@ import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
+import org.icatproject.ids.finiteStateMachine.FiniteStateMachine;
 import org.icatproject.ids.helpers.Constants;
 import org.icatproject.ids.helpers.ValueContainer;
 import org.icatproject.ids.v3.RequestHandlerService;
 import org.icatproject.ids.v3.ServiceProvider;
-import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
 
 @Path("/")
 @Stateless
