+++ src/main/java/org/icatproject/ids/requestHandlers/IsReadOnlyHandler.java
@@ -1,4 +1,4 @@
-package org.icatproject.ids.v3.handlers;
+package org.icatproject.ids.requestHandlers;
 
 import java.util.HashMap;
