+++ src/main/java/org/icatproject/ids/requestHandlers/PrepareDataHandler.java
@@ -1,16 +1,18 @@
-package org.icatproject.ids.v3.handlers;
+package org.icatproject.ids.requestHandlers;
 
+import java.io.BufferedOutputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.file.Files;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
 
 import org.icatproject.IcatException_Exception;
-import org.icatproject.ids.LockManager.Lock;
-import org.icatproject.ids.LockManager.LockType;
 import org.icatproject.ids.dataSelection.DataSelectionBase;
 import org.icatproject.ids.enums.CallType;
-import org.icatproject.ids.enums.DeferredOp;
 import org.icatproject.ids.enums.RequestType;
 import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.BadRequestException;
@@ -21,17 +23,16 @@ import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.helpers.ValueContainer;
 import org.icatproject.ids.models.DataInfoBase;
-import org.icatproject.ids.plugin.AlreadyLockedException;
 import org.icatproject.ids.v3.RequestHandlerBase;
 import org.icatproject.ids.v3.ServiceProvider;
 
 import jakarta.json.Json;
 import jakarta.json.stream.JsonGenerator;
 
-public class WriteHandler extends RequestHandlerBase {
+public class PrepareDataHandler extends RequestHandlerBase {
 
-    public WriteHandler() {
-        super(new StorageUnit[]{StorageUnit.DATAFILE, StorageUnit.DATASET, null}, RequestType.WRITE);
+    public PrepareDataHandler() {
+        super(new StorageUnit[]{StorageUnit.DATAFILE, StorageUnit.DATASET, null}, RequestType.PREPAREDATA);
     }
 
     @Override
@@ -39,6 +40,7 @@ public class WriteHandler extends RequestHandlerBase {
             throws BadRequestException, InternalException, InsufficientPrivilegesException, NotFoundException,
             DataNotOnlineException, NotImplementedException {
         
+                
         long start = System.currentTimeMillis();
         var serviceProvider = ServiceProvider.getInstance();
 
@@ -46,62 +48,58 @@ public class WriteHandler extends RequestHandlerBase {
         String investigationIds = parameters.get("investigationIds").getString();
         String datasetIds = parameters.get("datasetIds").getString();
         String datafileIds = parameters.get("datafileIds").getString();
+        boolean compress = parameters.get("compress").getBool();
+        boolean zip = parameters.get("zip").getBool();
         String ip = parameters.get("ip").getString();
 
         // Log and validate
-        logger.info("New webservice request: write " + "investigationIds='" + investigationIds + "' " + "datasetIds='"
-                + datasetIds + "' " + "datafileIds='" + datafileIds + "'");
-
-        if (!serviceProvider.getPropertyHandler().getEnableWrite()) {
-            throw new NotImplementedException("This operation has been configured to be unavailable");
-        }
+        logger.info("New webservice request: prepareData " + "investigationIds='" + investigationIds + "' "
+                + "datasetIds='" + datasetIds + "' " + "datafileIds='" + datafileIds + "' " + "compress='" + compress
+                + "' " + "zip='" + zip + "'");
 
         validateUUID("sessionId", sessionId);
 
-        final DataSelectionBase dataSelection = this.getDataSelection(sessionId, investigationIds, datasetIds, datafileIds);
+        final DataSelectionBase dataSelection = this.getDataSelection(sessionId,
+                investigationIds, datasetIds, datafileIds);
 
         // Do it
+        String preparedId = UUID.randomUUID().toString();
+
         Map<Long, DataInfoBase> dsInfos = dataSelection.getDsInfo();
+        Set<Long> emptyDs = dataSelection.getEmptyDatasets();
+        Map<Long, DataInfoBase> dfInfos = dataSelection.getDfInfo();
+        
+        dataSelection.restoreDataInfos();
 
-        try (Lock lock = serviceProvider.getLockManager().lock(dsInfos.values(), LockType.SHARED)) {
-            if (twoLevel) {
-                boolean maybeOffline = false;
-                for (DataInfoBase dataInfo : dataSelection.getPrimaryDataInfos().values()) {
-                    if (!dataSelection.existsInMainStorage(dataInfo)) {
-                        maybeOffline = true;
-                    }
-                }
-                if (maybeOffline) {
-                    throw new DataNotOnlineException("Requested data is not online, write request refused");
-                }
-            }
+        if (dataSelection.mustZip()) {
+            zip = true;
+        }
 
-            logger.info("### PreScheduleTask - StorageUnit: " + storageUnit);
-            dataSelection.scheduleTasks(DeferredOp.WRITE);
-            logger.info("### PostScheduleTask");
-
-        } catch (AlreadyLockedException e) {
-            logger.debug("Could not acquire lock, write failed");
-            throw new DataNotOnlineException("Data is busy");
+        logger.debug("Writing to " + preparedDir.resolve(preparedId));
+        try (OutputStream stream = new BufferedOutputStream(Files.newOutputStream(preparedDir.resolve(preparedId)))) {
+            pack(stream, zip, compress, dsInfos, dfInfos, emptyDs);
         } catch (IOException e) {
-            logger.error("I/O error " + e.getMessage() + " writing");
             throw new InternalException(e.getClass() + " " + e.getMessage());
         }
 
-        if (serviceProvider.getLogSet().contains(CallType.MIGRATE)) {
+        logger.debug("preparedId is " + preparedId);
+
+        if (serviceProvider.getLogSet().contains(CallType.PREPARE)) {
             try {
                 ByteArrayOutputStream baos = new ByteArrayOutputStream();
                 try (JsonGenerator gen = Json.createGenerator(baos).writeStartObject()) {
                     gen.write("userName", serviceProvider.getIcat().getUserName(sessionId));
                     addIds(gen, investigationIds, datasetIds, datafileIds);
+                    gen.write("preparedId", preparedId);
                     gen.writeEnd();
                 }
-                serviceProvider.getTransmitter().processMessage("write", ip, baos.toString(), start);
+                String body = baos.toString();
+                serviceProvider.getTransmitter().processMessage("prepareData", ip, body, start);
             } catch (IcatException_Exception e) {
                 logger.error("Failed to prepare jms message " + e.getClass() + " " + e.getMessage());
             }
         }
 
-        return ValueContainer.getVoid();
+        return new ValueContainer(preparedId);
     }
 }
