+++ src/main/java/org/icatproject/ids/thread/DsWriter.java
@@ -16,6 +16,7 @@ import org.icatproject.Datafile;
 import org.icatproject.Dataset;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.LockManager.Lock;
+import org.icatproject.ids.finiteStateMachine.FiniteStateMachine;
 import org.icatproject.ids.helpers.LocationHelper;
 import org.icatproject.ids.models.DataFileInfo;
 import org.icatproject.ids.models.DataSetInfo;
@@ -23,7 +24,6 @@ import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.plugin.ZipMapperInterface;
-import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
 
 /**
  * Copies dataset from main to archive
