+++ src/main/java/org/icatproject/ids/v3/RequestHandlerService.java
@@ -14,22 +14,22 @@ import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.helpers.ValueContainer;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
-import org.icatproject.ids.v3.handlers.ArchiveHandler;
-import org.icatproject.ids.v3.handlers.DeleteHandler;
-import org.icatproject.ids.v3.handlers.GetDataFileIdsHandler;
-import org.icatproject.ids.v3.handlers.GetDataHandler;
-import org.icatproject.ids.v3.handlers.GetIcatUrlHandler;
-import org.icatproject.ids.v3.handlers.GetServiceStatusHandler;
-import org.icatproject.ids.v3.handlers.GetSizeHandler;
-import org.icatproject.ids.v3.handlers.GetStatusHandler;
-import org.icatproject.ids.v3.handlers.IsPreparedHandler;
-import org.icatproject.ids.v3.handlers.IsReadOnlyHandler;
-import org.icatproject.ids.v3.handlers.IsTwoLevelHandler;
-import org.icatproject.ids.v3.handlers.PrepareDataHandler;
-import org.icatproject.ids.v3.handlers.PutHandler;
-import org.icatproject.ids.v3.handlers.ResetHandler;
-import org.icatproject.ids.v3.handlers.RestoreHandler;
-import org.icatproject.ids.v3.handlers.WriteHandler;
+import org.icatproject.ids.requestHandlers.ArchiveHandler;
+import org.icatproject.ids.requestHandlers.DeleteHandler;
+import org.icatproject.ids.requestHandlers.GetDataFileIdsHandler;
+import org.icatproject.ids.requestHandlers.GetDataHandler;
+import org.icatproject.ids.requestHandlers.GetIcatUrlHandler;
+import org.icatproject.ids.requestHandlers.GetServiceStatusHandler;
+import org.icatproject.ids.requestHandlers.GetSizeHandler;
+import org.icatproject.ids.requestHandlers.GetStatusHandler;
+import org.icatproject.ids.requestHandlers.IsPreparedHandler;
+import org.icatproject.ids.requestHandlers.IsReadOnlyHandler;
+import org.icatproject.ids.requestHandlers.IsTwoLevelHandler;
+import org.icatproject.ids.requestHandlers.PrepareDataHandler;
+import org.icatproject.ids.requestHandlers.PutHandler;
+import org.icatproject.ids.requestHandlers.ResetHandler;
+import org.icatproject.ids.requestHandlers.RestoreHandler;
+import org.icatproject.ids.requestHandlers.WriteHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
